1. Removed everything related to email templates. 2. Fixed history bug for pages found
