--- app/Events/Backend/EmailTemplates/EmailTemplateDeleted.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace App\Events\Backend\EmailTemplates;
-
-use Illuminate\Queue\SerializesModels;
-
-/**
- * Class EmailTemplateDeleted.
- */
-class EmailTemplateDeleted
-{
-    use SerializesModels;
-
-    /**
-     * @var
-     */
-    public $emailtemplates;
-
-    /**
-     * @param $emailtemplates
-     */
-    public function __construct($emailtemplates)
-    {
-        $this->emailtemplates = $emailtemplates;
-    }
-}
