--- app/Models/EmailTemplates/Traits/Attribute/EmailTemplateAttribute.php
@@ -1,37 +1,0 @@
-<?php
-
-namespace App\Models\EmailTemplates\Traits\Attribute;
-
-/**
- * Class EmailTemplateAttribute.
- */
-trait EmailTemplateAttribute
-{
-    /**
-     * @return string
-     */
-    public function getActionButtonsAttribute()
-    {
-        return '<div class="btn-group action-btn">'.$this->getEditButtonAttribute('edit-email-template', 'admin.emailtemplates.edit').'</div>';
-    }
-
-    /**
-     * @return string
-     */
-    public function getStatusLabelAttribute()
-    {
-        if ($this->isActive()) {
-            return "<label class='label label-success'>".trans('labels.general.active').'</label>';
-        }
-
-        return "<label class='label label-danger'>".trans('labels.general.inactive').'</label>';
-    }
-
-    /**
-     * @return bool
-     */
-    public function isActive()
-    {
-        return $this->status == 1;
-    }
-}
