+++ app/Providers/EventServiceProvider.php
@@ -41,8 +41,7 @@ class EventServiceProvider extends ServiceProvider
         \App\Listeners\Backend\Access\User\UserEventListener::class,
         \App\Listeners\Backend\Access\Role\RoleEventListener::class,
         \App\Listeners\Backend\Access\Permission\PermissionEventListener::class,
-        \App\Listeners\Backend\CMSPages\CMSPageEventListener::class,
-        \App\Listeners\Backend\EmailTemplates\EmailTemplateEventListener::class,
+        \App\Listeners\Backend\Pages\PageEventListener::class,
         \App\Listeners\Backend\BlogCategories\BlogCategoryEventListener::class,
         \App\Listeners\Backend\BlogTags\BlogTagEventListener::class,
         \App\Listeners\Backend\Blogs\BlogEventListener::class,
