+++ database/seeds/DatabaseSeeder.php
@@ -19,9 +19,6 @@ class DatabaseSeeder extends Seeder
 
         $this->call(AccessTableSeeder::class);
         $this->call(HistoryTypeTableSeeder::class);
-        $this->call(EmailTemplateTypeTableSeeder::class);
-        $this->call(EmailTemplatePlaceholderTableSeeder::class);
-        $this->call(EmailTemplateTableSeeder::class);
         $this->call(SettingsTableSeeder::class);
         $this->call(PagesTableSeeder::class);
         $this->call(MenuTableSeeder::class);
