--- database/seeds/EmailTemplatePlaceholderTableSeeder.php
@@ -1,80 +1,0 @@
-<?php
-
-use Carbon\Carbon as Carbon;
-use Illuminate\Database\Seeder;
-use Illuminate\Support\Facades\DB;
-
-class EmailTemplatePlaceholderTableSeeder extends Seeder
-{
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
-    public function run()
-    {
-        if (env('DB_CONNECTION') == 'mysql') {
-            DB::table(config('module.email_templates.placeholders_table'))->truncate();
-        }
-
-        $data = [
-            [
-                'name'       => 'app_name',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'name',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'email',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'password',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'contact-details',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'confirmation_link',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'password_reset_link',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'header_logo',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'footer_logo',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'unscribe_link',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-            [
-                'name'       => 'status',
-                'created_at' => Carbon::now(),
-                'updated_at' => Carbon::now(),
-            ],
-        ];
-
-        DB::table(config('module.email_templates.placeholders_table'))->insert($data);
-    }
-}
